v0.2.0 Adding drawing capabilites
After the reset of how the map will be drawn, a lot of the methods have been removed as they are unnecessary, or not worked on yet.

Currently created a handler to use with osmium so that nodes on the osm map can be extracted allowing for the map to be drawn.